gcc: prevent crash on Eclair and older platforms.
The point of this patch is to work-around a bug in the Eclair
dynamic linker, which doesn't support weak symbols. By default,
libsupc++ and libstdc++ generate static C++ constructors that
reference weak symbols.

When they are statically linked into shared libraries, the
corresponding code is referenced in its .init_array section
and run when the shared library is loaded.

On Eclair and previous release, the weak symbol is not resolved
before the constructor are launched, resulting in a crash when
the PLT entry tries to jump to address 0.

By not generating weak symbol references, we avoid the problem
completely. And we don't need them because the pthread symbols
are all in the C library on Android, unlike legacy Linux systems
which put them in libpthread.so (and provide weak stubs in libc.so).

Change-Id: I289db08a9ea567f0574bb2058516bfeb01c68a2c